Pull nested classes out of SnowballScene
