+++ oprheum-game/SnowballScene.swift
@@ -18,128 +18,6 @@ func limit(val: CGFloat) -> (CGFloat) {
     return result
 }
 
-class SnowballNode: SKNode {
-    var snowballNode: SKShapeNode!
-    var mass: CGFloat = 8
-    var mass0: CGFloat = 0
-    let whiteColor = UIColor(red: 1.00, green: 1.00, blue: 1.00, alpha: 1.00)
-
-    
-    override init() {
-        super.init()
-        
-        
-        self.name = "Snowball"
-        redrawSnowball()
-        unselect()
-    }
-    
-    func setSnowballMass(mass: CGFloat) {
-        self.mass = mass
-        redrawSnowball()
-    }
-    
-    func redrawSnowball() {
-        self.removeAllChildren()
-        let radius = mass * 10
-        snowballNode = SKShapeNode(circleOfRadius: radius)
-        self.physicsBody = SKPhysicsBody(circleOfRadius: radius)
-        self.physicsBody?.mass = mass
-        self.addChild(snowballNode)
-        snowballNode.fillColor = whiteColor
-    }
-    
-    func pinchBegan(scale: CGFloat){
-        mass0 = mass
-    }
-    
-    func pinchChanged(scale: CGFloat) {
-        setSnowballMass(mass0 * scale)
-    }
-    
-    func pinchEnded(scale: CGFloat) {
-        setSnowballMass(mass0 * scale)
-    }
-    
-    func select() {
-        snowballNode.fillColor = whiteColor
-        // snowballNode.fillColor = SKColor.whiteColor()
-    }
-    
-    func unselect() {
-        snowballNode.fillColor = SKColor(red: 0.621, green: 0.864, blue: 1.000, alpha: 1.000)
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-class RampNode: SKNode {
-    var rampNode: SKShapeNode!
-    var triangleHeight: CGFloat = 400
-    var triangleWidth: CGFloat = 400
-    var x0: CGFloat = 0
-    var y0: CGFloat = 0
-    
-    override init() {
-        super.init()
-        
-        self.name = "Ramp"
-        
-        rampNode = SKShapeNode()
-        rampNode.physicsBody?.mass=10000
-        self.redrawTriangle(triangleWidth, height: triangleHeight)
-        self.unselect()
-        
-        self.addChild(rampNode)
-    }
-
-    func redrawTriangle(width: CGFloat, height: CGFloat) {
-        let rampPath = CGPathCreateMutable()
-        CGPathMoveToPoint(rampPath, nil, 0, height)
-        CGPathAddLineToPoint(rampPath, nil, 0, 0)
-        CGPathAddLineToPoint(rampPath, nil, width, 0)
-        CGPathCloseSubpath(rampPath)
-        
-        rampNode.path = rampPath
-        rampNode.physicsBody = SKPhysicsBody(polygonFromPath: rampPath)
-    }
-    
-    func pinchBegan(touch1:CGPoint, touch2:CGPoint){
-        x0 = touch2.x - touch1.x
-        y0 = touch1.y - touch2.y
-    }
-    
-    func pinchChanged(touch1:CGPoint, touch2:CGPoint) {
-        let xScale = (touch2.x - touch1.x) / x0
-        let yScale = (touch1.y - touch2.y) / y0
-        let xLimit = limit(triangleWidth * xScale)
-        let yLimit = limit(triangleHeight * yScale)
-        redrawTriangle(xLimit, height: yLimit)
-    }
-    
-    func pinchEnded(touch1:CGPoint, touch2:CGPoint) {
-        let xScale = (touch2.x - touch1.x) / x0
-        let yScale = (touch1.y - touch2.y) / y0
-        triangleWidth = limit(triangleWidth * xScale)
-        triangleHeight = limit(triangleHeight * yScale)
-        redrawTriangle(triangleWidth, height: triangleHeight)
-    }
-    
-    func select() {
-        rampNode.fillColor = SKColor.redColor()
-    }
-    
-    func unselect() {
-        rampNode.fillColor = SKColor(red: 0.039, green: 1.000, blue: 0.004, alpha: 1.00)
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
 class SnowballScene: SKScene {
     var gameViewController: GameViewController!
     var difficulty: CGFloat = 0.5
